requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip==0.1.0
cmake==3.29.3
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
einops==0.3.0
executing==2.0.1
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
future==1.0.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.64.0
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipdb==0.13.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpickle==1.5.2
kiwisolver==1.4.5
lightning-utilities==0.11.2
lit==18.1.6
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.5
munch==2.5.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==24.0
pandas==1.1.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==8.2.0
prompt_toolkit==3.0.44
protobuf==5.27.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==2.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.1.4
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sacred==0.8.2
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
timm==1.0.3
tokenizers==0.9.4
torch==2.0.1
torchmetrics==1.4.0.post0
torchvision==0.15.2
tqdm==4.56.0
traitlets==5.14.3
transformers==4.2.1
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
urllib3==2.2.1
utils==1.0.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.2
